Move legacy tests to test/legacy
